use async-sqlite instead of self impl
